- Added javadoc.
- Added more information to the readme.md.

- Added so the annotation logger can measure the instant deletion of a &line and &block annotation by taking the highlighting as its start time.

- Added so the annotation logger can take the highlighting as its starting time when instantly creating an annotation with "Surround with Feature annotation". There is a bug here that I have written below.

(OBS: However there is a bug where if you highlight a text and then right click outside the highlight text and instantly create an annotation with "Surround with Feature annotation", it will still take the highlight as the starting time. There is a way to fix this where you only take the highlight time if the rightclick is inside the lines of the highlight, I will be looking into this more shortly).